% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{erpSubjectAveraged}
\alias{erpSubjectAveraged}
\title{Subject-Averaged Data}
\format{A data.frame with 3,328 rows and 5 variables:
\describe{
  \item{value}{Average ERP amplitude 250-700ms post-feedback onset, averaged across epochs}
  \item{ERPset}{subject id}
  \item{chindex}{numerical label for each electrode}
  \item{chlabel}{abbreviation for electrode}
  \item{bini}{tot condition; 1 = tot, 2 = no-tot}

}}
\source{
\url{https://github.com/pab2163/TOT_ERP}
}
\usage{
erpSubjectAveraged
}
\description{
A data set with averaged subject ERP data from the TOT ERP study, directly from ERPLAB preprocessing output (http://www.erpinfo.org/erplab.html)
}
\keyword{datasets}
